pyproject.toml
[tool.poetry]
name = "shakugenerator"
version = "0.1.0"
description = ""
authors = ["ElectricShakuhachi <haavistoaaro@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
invoke = "^1.6.0"
MIDIUtil = "^1.2.1"
python-dotenv = "^0.19.2"
midi2audio = "^0.1.1"
fluidsynth = "^0.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
coverage = "^6.2"
pylint = "^2.12.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
